+++ src/hash/hash.js
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import { resolve } from 'path';
 import crypto from 'crypto';
-import { currentlyDirectory } from '../os/currently_directory.js';
+import { currentlyDirectory } from '../os/currentlyDirectory.js';
 import { finished } from 'stream';
 
 export const calculateHash = async (pathToFile) => {
