+++ src/os/architecture.js
@@ -1,5 +1,5 @@
 import os from 'os';
-import { currentlyDirectory } from './currently_directory.js';
+import { currentlyDirectory } from './currentlyDirectory.js';
 
 export const showArchitecture = () => {
   try {
